+++ package.json
@@ -4,10 +4,10 @@
 	"license": "MIT",
 	"description": "A powerful \"short-cutter\" to your console for you and your team!",
 	"scripts": {
-		"test": "mocha lib/**/*.spec.js -R nyan",
+		"test": "mocha test/**/*.spec.js -R nyan",
 		"watch": "npm test --watch",
 		"lint": "eslint \"bin/e.js\" \"lib/**/*.js\" --fix",
-		"test-ci": "mocha lib/**/*.spec.js -R min",
+		"test-ci": "mocha test/**/*.spec.js -R min",
 		"test-docker": "echo removeme",
 		"test-docker-temp": "mocha test/docker.js -R min --timeout 150000",
 		"coverage": "nyc --reporter=html --reporter=text npm run test-ci --timeout=3000",
@@ -38,7 +38,6 @@
 			"lib/**/*.js"
 		],
 		"exclude": [
-			"lib/**/*.spec.js"
 		],
 		"sourceMap": false,
 		"instrument": true,
